refactor(iocuak): extract classMetadata and reflectMetadata modules
