+++ packages/iocuak/src/classMetadata/models/domain/ClassMetadata.ts
