+++ packages/iocuak/src/e2e/definitions/container/parameters/containerModuleMetadata/getContainerModuleMetadataWithContainerModuleFactoryMetadataImports.ts
@@ -1,10 +1,10 @@
 import sinon from 'sinon';
 
+import { inject } from '../../../../../classMetadata/decorators/inject';
 import { ContainerModuleBindingServiceApi } from '../../../../../container/services/api/ContainerModuleBindingServiceApi';
 import { ContainerModuleApi } from '../../../../../containerModule/models/api/ContainerModuleApi';
 import { ContainerModuleClassMetadataApi } from '../../../../../containerModuleMetadata/models/api/ContainerModuleClassMetadataApi';
 import { ContainerModuleFactoryMetadataApi } from '../../../../../containerModuleMetadata/models/api/ContainerModuleFactoryMetadataApi';
-import { inject } from '../../../../../metadata/decorators/inject';
 import { injectable } from '../../../../../metadata/decorators/injectable';
 import { getTypeServiceWithNoDependenciesParameter } from '../../../common/parameters/typeService/getTypeServiceWithNoDependenciesParameter';
 import { TypeServiceParameter } from '../../../common/parameters/typeService/TypeServiceParameter';
