+++ packages/iocuak/src/index.ts
@@ -1,3 +1,7 @@
+import { inject } from './classMetadata/decorators/inject';
+import { injectFrom } from './classMetadata/decorators/injectFrom';
+import { injectFromBase } from './classMetadata/decorators/injectFromBase';
+import { ClassMetadataApi } from './classMetadata/models/api/ClassMetadataApi';
 import { Newable } from './common/models/domain/Newable';
 import { ServiceId } from './common/models/domain/ServiceId';
 import { ContainerApi } from './container/modules/api/ContainerApi';
@@ -5,14 +9,10 @@ import { ContainerModuleBindingServiceApi } from './container/services/api/Conta
 import { ContainerServiceApi } from './container/services/api/ContainerServiceApi';
 import { ContainerModuleApi } from './containerModule/models/api/ContainerModuleApi';
 import { ContainerModuleMetadataApi } from './containerModuleMetadata/models/api/ContainerModuleMetadataApi';
-import { inject } from './metadata/decorators/inject';
 import { injectable } from './metadata/decorators/injectable';
-import { injectFrom } from './metadata/decorators/injectFrom';
-import { injectFromBase } from './metadata/decorators/injectFromBase';
 import { BindingApi } from './metadata/models/api/BindingApi';
 import { BindingScopeApi } from './metadata/models/api/BindingScopeApi';
 import { BindingTypeApi } from './metadata/models/api/BindingTypeApi';
-import { ClassMetadataApi } from './metadata/models/api/ClassMetadataApi';
 import { InjectableOptionsApi } from './metadata/models/api/InjectableOptionsApi';
 import { TypeBindingApi } from './metadata/models/api/TypeBindingApi';
 import { ValueBindingApi } from './metadata/models/api/ValueBindingApi';
