+++ packages/iocuak/src/metadata/services/api/MetadataServiceApi.ts
@@ -1,5 +1,5 @@
+import { ClassMetadataApi } from '../../../classMetadata/models/api/ClassMetadataApi';
 import { Newable } from '../../../common/models/domain/Newable';
-import { ClassMetadataApi } from '../../models/api/ClassMetadataApi';
 import { TypeBindingApi } from '../../models/api/TypeBindingApi';
 
 export interface MetadataServiceApi {
